Melhorias e implementações
Adicionamos o suporte à DLL ISBNQuery para recuperar informações online de ISBN10, ISBN13, autores, títulos, etc. Também melhoramos a forma que os dados na saída de um livro são registrados. Ainda, adicionamos um formulário para registro de livros de leitura. Foi criado também um novo método de Insert para registro no banco de dados. O novo método executa a rotina de DataInsert de forma mais segura. Recomenda-se de forma progressiva a migração do método DatabaseController.DML para o DatabaseController.InsertData.

Por fim, atualizamos alguns pacotes Nuget do EasyLi.

+++ biblioteca/Forms/CorrecaoLivros.cs
@@ -78,6 +78,9 @@ namespace biblioteca
 
         private void ExcluirClick(object sender, EventArgs e)
         {
+            if (Livros.SelectedRows.Count < 1)
+                return;
+
             if (DialogResult.Yes == MessageBox.Show("Você tem certeza que deseja apagar este registros?", "Apagar livro", MessageBoxButtons.YesNo, MessageBoxIcon.Question))
             {
                 DatabaseController.DML(String.Format("delete from tb_livros where id = '{0}'", Livros.SelectedRows[0].Cells[0].Value.ToString()));
